% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj2rankadj.R
\name{adj2rankadj}
\alias{adj2rankadj}
\title{Convert network matrix to network rank matrix}
\usage{
adj2rankadj(adjmatrix, directed = F)
}
\arguments{
\item{adjmatrix}{A network matrix}

\item{directed}{Logical. If TRUE, the network is considered as directed. If FALSE, the upper triangular part of the matrix is used to calcuate the rank matrix}
}
\value{
an network connection rank matrix
}
\description{
Convert network matrix to network rank matrix
}

