% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmatrix2edgelist.R
\name{adjmatrix2edgelist}
\alias{adjmatrix2edgelist}
\title{Convert network matrix to edge list}
\usage{
adjmatrix2edgelist(adjmatrix, cutoff = 0, directed = F)
}
\arguments{
\item{adjmatrix}{a network matrix}

\item{cutoff}{threshod to cut the edge list}

\item{directed}{logical, if \code{FLASE} the adjmatrix is transformed to symmetric matrix}
}
\value{
a edge list
}
\description{
The inferred network matrix is converted to edge list
}
\examples{
##load data
library(networkBMA)
library(RLowPCor)
data(dream4)
data.exp<-dream4ts10[[1]][,-c(1:2)]
genes<-colnames(data.exp)
##build correlation network
inf.cor<-abs(cor(data.exp))
diag(inf.cor)<-0
##convert matrix to edge list
adjmatrix2edgelist(inf.cor)
}

