% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.consensus.R
\name{average.consensus}
\alias{average.consensus}
\title{Consensus network from average rank}
\usage{
average.consensus(adjmatrix.list, directed = F)
}
\arguments{
\item{adjmatrix.list}{A list of network prediciton matrices with same rownames and colnames.}

\item{directed}{Logical. If TRUE, the networkis considered as directed. If FALSE, the upper triangular part of the matrix is used to calcuate the rank matrix}
}
\value{
a network with rank weighted edges. The weights are rescale to 0-1 and hihger values indicate higher ranks.
}
\description{
Consensus network is built of taking the average ranks of the edges from multiple network predictions.
}
\examples{
##create two random networks
library(RLowPCor)
set.seed(4)
net1<-abs(matrix(rnorm(16),4,4))
net1<-pmax(net1,t(net1))
diag(net1)<-0
set.seed(5)
net2<-abs(matrix(rnorm(16),4,4))
net2<-pmax(net2,t(net2))
diag(net2)<-0
dimnames(net1)<-dimnames(net2)<-list(letters[1:4],letters[1:4])
net.list<-list(net1=net1,net2=net2)
inf.consensus<-average.consensus(adjmatrix.list = net.list,directed = F)
adj2rankadj(net1)
adj2rankadj(net2)
inf.consensus
}

