% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2mi.R
\name{cor2mi}
\alias{cor2mi}
\title{Convert correlation matrix to MI matrix}
\usage{
cor2mi(x)
}
\arguments{
\item{x}{a correlation matrix}
}
\value{
\code{cor2mi} returns a mutual information matrix.
}
\description{
The correlation \eqn{\rho_{ij}} between node \eqn{i} and \eqn{j} is converted into mutual information (MI) matrix by using equation [1]:
\deqn{MI=-\frac{1}{2} (1-\rho_{ij}^2)}
if the observation data is normal distributed.
}
\references{
[1] Meyer PE, Kontos K, Lafitte F, Bontempi G: Information-theoretic inference of large transcriptional
regulatory networks. EURASIP J Bioinform Syst Biol 2007:79879.
}

