% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2statistics.R
\name{cor2statistics}
\alias{cor2statistics}
\title{Convert correlation to statistical significance measures.}
\usage{
cor2statistics(cor.vector, ...)
}
\arguments{
\item{cor.vector}{a correlation vector}

\item{...}{parameters used in \code{\link{fdrtool}}.}
}
\value{
\code{cor2statistics} returns a list includes correlation matrix and p-value matrix.
}
\description{
\code{cor2statistics} convert a correlation vector to p-values, false discovery rates (FDRs) and probability of connections.
}
\seealso{
\code{\link{fdrtool}}
}

