% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist2adjmatrix.R
\name{edgelist2adjmatrix}
\alias{edgelist2adjmatrix}
\title{Convert edge list to network matrix}
\usage{
edgelist2adjmatrix(edgelist, genes, cutoff = 0, directed = F)
}
\arguments{
\item{edgelist}{a edge list of network}

\item{genes}{gene names}

\item{cutoff}{the threshold to cut the edge list}

\item{directed}{logical, to create directed or undirected network matrix}
}
\value{
a network matrix
}
\description{
The function is to convert edge list to network matrix.
}
\examples{
 library(networkBMA)
 library(RLowPCor)
 ##load DREAM4 size100_1 datasets
 data(dream4)
 data.exp<-dream4ts10[[1]][,-c(1:2)]
 genes<-colnames(data.exp)
 ref.edge<-dream4gold10[[1]]
 ref.adj<-edgelist2adjmatrix(ref.edge,genes)
}

