% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstPC.R
\name{firstPC}
\alias{firstPC}
\title{First order partial correlation}
\usage{
firstPC(data.exp, edgelist, controlist = NULL, method = "pearson",
  progressbar = T)
}
\arguments{
\item{data.exp}{gene expression data matrix with genes in columns and samples in rows.}

\item{edgelist}{an edge list for fisrt order PC calculation. The PC values are calcuated over each pair of genes in the gene list.}

\item{controlist}{list of vectors of neighbour genes for each pair of genes in \code{edgelist}. The vector order in the list must match to the order of corresponding paired genes in the
input \code{edgelist}. For example the i element in \code{controlist} contains the shared neighbours of gene \code{edgelist[i,1]} and gene \code{edgelist[i,2]}.}

\item{method}{estimators of correlation. Options are "pearson", "spearman" and "kendall".}

\item{logical.}{If TRUE, a progressbar will show to indicate the code runing percentage.}
}
\value{
\code{firstPC} returns a data frame with columns of regulators, target genes, edge correlation weigthes, p-values, FRD and connection probability.
}
\description{
First order PC is calcuated according to the edge connections in the input edgelist. The correlation of a pair of genes are calcuated as (a) correlation if they do not connect
to the same neighbour genes and (b) the maximum of first order PC values by removing the controls one by one if they connect to shared neighbour genes. After that, the
\code{\link{cor2statistics}} is used to calculate the p-value, FDR and connection probability from the output correlation.
}
\seealso{
\code{\link{cor2statistics}}, \code{\link{RLowPC}}
}

