% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowPC.R
\name{LowPC}
\alias{LowPC}
\title{Low order partial correlation}
\usage{
LowPC(data.exp, cutoff = 0.05, cutat = "pval", method = "pearson",
  progressbar = T)
}
\arguments{
\item{data.exp}{gene expression data matrix with genes in columns and samples in rows.}

\item{cutoff}{a cut-off of p-values or conenction proabibility.}

\item{cutat}{a string indicates to cut the results at p-value "pval" or probability "prob" with a threshold \code{cutoff}.}

\item{method}{a string character of method used to estimate correlation. Options are "pearson", "spearman" and "kendall".}

\item{progressbar}{logical. If TRUE, a progressbar will show to indicate the code runing percentage.}
}
\value{
\code{LowPC} returns a list includes the results of zero order PC, first order PC and second order PC.
}
\description{
Calculate the low order PC as shown in paper [1].
}
\details{
Zero order PC (correlation) is firstly computed. The significant edges in zero order PC network are
used to calcuate first order PC. After that, the significant edges in first order PC are used to calculate second order PC. Instead of regressing all the remained genes one by one or
pair by pair, the calculation of first and second PC only based on the subset of edges that significantly connect to the genes, which greatly reduces the computational cost compared to the
paper [1]. For example, in a network with \eqn{n} genes, the first order PC between gene \eqn{i} and \eqn{j} is calculated by removing \eqn{m} shared genes, which significantly connect to \eqn{i} and
\eqn{j}, one by one rather than all the \eqn{n-2} remained genes (\eqn{m<n-2}). If a pair of genes do not connect to the same set of genes,
the correlation between them is represeted as correlation, while if they connected only one shared gene, the edge connection is weighted as
first order PC. The correlation significance is estimated using R package \code{\link{fdrtool}}.
}
\references{
1. Zuo Y, Yu G, Tadesse MG, Ressom HW: Biological network inference using low order partial correlation. Methods (San Diego, Calif) 2014, 69(3):266-273.
}
\seealso{
\code{\link{zeroPC}}, \code{\link{firstPC}}, \code{\link{secondPC}}, \code{\link{RLowPC}} and \code{\link{cor2statistics}}.
}

