% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.degree.R
\name{node.degree}
\alias{node.degree}
\title{Calculate degree of nodes in gene networks}
\usage{
node.degree(net, directed = F)
}
\arguments{
\item{net}{input network matrix or edge list with three columns, i.e. regulators, targets and weights.}

\item{directed}{logical, if \code{FALSE}, the input network is treated as undirected.}
}
\value{
a vector of node degrees
}
\description{
Calculate degree of nodes in gene networks
}

