% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAUC.R
\name{plotAUC}
\alias{plotAUC}
\title{Plot AUC curves}
\usage{
plotAUC(table.methods, plot.method = c("auroc", "aupr"), roc.lim = c(0, 1,
  0, 1), pr.lim = c(0, 1, 0, 1), fill = F, color = NA, ...)
}
\arguments{
\item{table.methods}{a confusion table list of multiple methods, which are from the output of \code{\link{table.evalution}}. The names of the list will be used
to plot the legend of the curves.}

\item{plot.method}{a character string indicates to plot "auroc" or "aupr" curves.}

\item{roc.lim}{a vector to tell the plot limits, i.e. xmin,xmax,ymin,ymax, of AUROC curves.}

\item{pr.lim}{a vector to tell the plot limits, i.e. xmin,xmax,ymin,ymax, of AUPR curves.}

\item{fill}{logical. If TRUE, the area under the curvers will be filled with colors.}
}
\value{
Plots.
}
\description{
The function is used to plot AUROC, AUPR, pAUROC and pAUPR curves.
}
\details{
The fpr or recall cutoffs to make pAUROC or pAUPR curves can be generate from the funciton \code{\link{confusion}} and specify in \code{roc.lim} and \code{pr.lim}
}
\examples{
##load librarys
library(networkBMA)
library(RLowPCor)
library(minet)
library(gridEXtra)
library(gglot2)
##get dream4 datasets
data(dream4)
data.exp<-dream4ts100[[1]][,-c(1:2)]
genes<-colnames(data.exp)
ref.edge<-dream4gold100[[1]]
ref.adj<-edgelist2adjmatrix(ref.edge,genes)
##infer gene networks
inf.cor<-abs(cor(data.exp))
diag(inf.cor)<-0
inf.mi<-build.mim(data.exp)
inf.clr<-clr(inf.mi)
inf.mrnet<-mrnet(inf.mi)
##generate confusion tables
table.cor<-table.evaluate(inf.adj = inf.cor,ref.adj = ref.adj)
table.mi<-table.evaluate(inf.adj=inf.mi,ref.adj = ref.adj)
table.clr<-table.evaluate(inf.adj=inf.clr,ref.adj = ref.adj)
table.mrnet<-table.evaluate(inf.adj=inf.mrnet,ref.adj = ref.adj)
##put confusion tables into list, and set names as the methods
table.methods<-list(cor=table.cor,mi=table.mi,clr=table.clr,mrnet=table.mrnet)
plotAUC(table.methods,plot.method = 'auroc',roc.lim = c(0,0.5,0,1))

}

