% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.net.R
\name{random.net}
\alias{random.net}
\title{Generate random network}
\usage{
random.net(input, nblock, from = "data", directed = F)
}
\arguments{
\item{input}{a gene expression \code{dataframe} with variables in columns and samples in rows or a vector of node names.}

\item{nblock}{a numeric number to indicate how many blocks that consist the datasets. For example, to tell how many replicates or experiments are in the datasets.
The datasets to generate random network will be resorted in block-wise.}

\item{from}{a character string to tell the random network generation is based on nodes size or dataset.Options are "node" and "data".}

\item{directed}{logical, if \code{FALSE}, the random network generated from nodes will be converted to symmetric matrix.}
}
\value{
The resorted datasets if \code{from='data'} or a random network matrix, the edge weights of which are in \eqn{[0,1]} if \code{from='node'}.
}
\description{
Random networks are generated from node size or from resort odering of samples in gene expression dataset.
}

