% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.zero.R
\name{remove.zero}
\alias{remove.zero}
\title{Remove the zero rows and columns of data frame}
\usage{
remove.zero(adjmatrix)
}
\arguments{
\item{adjmatrix}{a network matrix}
}
\value{
A new network matrix
}
\description{
The function is used to remove the zero rows and zero columns in network matrix
}

