% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondPC.R
\name{secondPC}
\alias{secondPC}
\title{Second order partial correlation}
\usage{
secondPC(data.exp, edgelist, controlist = NULL, method = "pearson",
  progressbar = T)
}
\arguments{
\item{data.exp}{gene expression data matrix with genes in columns and samples in rows.}

\item{edgelist}{an edge list for second order PC calculation. The PC values are calcuated over each pair of genes in the gene list.}

\item{controlist}{list of vectors of neighbour genes for each pair of genes in \code{edgelist}. The vector order in the list must match to the order of corresponding paired genes in the
input \code{edgelist}. For example the i element in \code{controlist} contains the shared neighbours of gene \code{edgelist[i,1]} and gene \code{edgelist[i,2]}.
the i}

\item{method}{estimators of correlation. Options are "pearson", "spearman" and "kendall".}

\item{logical.}{If TRUE, a progressbar will show to indicate the code runing percentage.}
}
\value{
\code{secondPC} returns a data frame with columns of regulators, target genes, edge correlation weigthes, p-values, FRD and connection probability.
}
\description{
Second order PC is calcuated according to the edge connections in the input edgelist. The correlation of a pair of genes are calcuated as (a) correlation if they do not connect
to the same neighbour genes; (b) first order PC if they only connect to one shared genes and (c) second order PC by removing the controls pair by pair if they connect
to more than two shared neighbours genes. After that, the \code{\link{fdrtool}} is used to calculate the p-value, FDR and connection probability from the correlation.
}

