% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondPCor.R
\name{secondPCor}
\alias{secondPCor}
\title{Second order partial correlation}
\usage{
secondPCor(adjmatrix, data.exp, method = "pearson", p.quant.prob = 1,
  progressbar = T)
}
\arguments{
\item{adjmatrix}{an adjacency matrix with 0 and 1 numerical values to indicate edge connections.}

\item{data.exp}{gene expression data matrix with variables in columns and samples in rows.}

\item{method}{estimators of correlation. Options are "pearson" and "spearman"}

\item{p.quant.prob}{distribution probability of p-value quantile cut-off. It is used to determine which p-value of all the range of p-values for a pair of nodes is selected to determine
the edge correlation.}
}
\value{
A list includes 3 \code{dataframe} of results. \code{cor} is the results of first order PC assigned to the input adjacency matrix. \code{pval} is the
p-value results and \code{adj} is the input adjacency matrix.
}
\description{
Second order PC is calcuated according to the edge connections in the input adjacency matrix. The adjacency matrix can be a fully connected network (e.g. correlation matrix)
or a sparse network (e.g. correlation matrix after applying a edge cut-off). The correlation and p-value of a pair of nodes are measurd as (1) if they have at least 2
shared neighbours, a set of second order PC values and their p-values are calculated by removing all the paired combinations of neighbours. Edge correlation is deterimed by the
\code{quantile(p.values, probs=p.quant.prob)} p-value and the corresponding PC value. Setting \code{probs=1} indicates the maximum p-value will be selected as the significance metric.
(2) if they only have one shared neighbour, first order PC and the corresponding p-value is calculated as in \code{\link{firstPCor}} and (3) if they do not connect to the same nodes, the connection and p-value are calculated from the zero order PC (the general correlation).
}

