% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.network.R
\name{top.network}
\alias{top.network}
\title{Select sub-network with top weigthed edges}
\usage{
top.network(net, top = 1000)
}
\arguments{
\item{net}{a netork matrix or edge list}

\item{top}{a numeric number to indicate the top weigthed edges, e.g. \code{top=1000}}
}
\value{
A subset of network matrix or edge list
}
\description{
Select sub-network with top weigthed edges
}

