% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroPC.R
\name{zeroPC}
\alias{zeroPC}
\title{Zero order partial correlation}
\usage{
zeroPC(data.exp, method = "pearson")
}
\arguments{
\item{data.exp}{gene expression data.}

\item{method}{"pearson", "spearman" or "kendall" method is used to calcuate the correlation values.}
}
\value{
\code{zeroPC} returns a data frame with columns of regulators, target genes, edge correlation weigthes,
p-values, FRD and connection probability.
}
\description{
The correlation and p-value of correlation.
}
\examples{
##load data
data(gnwdata)
##ts expression
data.exp<-gnwdata$size100$ts1[,-c(1:3)]
inf.zeroPC<-zeroPC(data.exp)
head(inf.zeroPC)
##refernece network

}

