% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmatrix2edgelist.R
\name{adjmatrix2edgelist}
\alias{adjmatrix2edgelist}
\title{Convert network matrix to edge list}
\usage{
adjmatrix2edgelist(adjmatrix, cutoff = 0, directed = F, order = T)
}
\arguments{
\item{adjmatrix}{a network matrix}

\item{cutoff}{threshod to cut the edge weights. Only the edges that pass the threshold will be shown in the output results.}

\item{directed}{logical, if \code{FLASE} the adjmatrix is transformed to symmetric matrix and the upper triangular part of the matrix is used to generate the edge list.}

\item{order}{logical, to order the edgelist according the edge weightes or not.}
}
\value{
\code{adjmatrix2edgelist} returns a data frame of edge list.
}
\description{
Convert the network matrix to edge list, in which the first column contains the regulators, the second column presents the target genes and edge weights are in the third column.
}
\examples{
##load data
library(RLowPC)
data(gnwdata)
data.exp<-gnwdata$size100$ts1[,-c(1:3)]
genes<-colnames(data.exp)
##build correlation network
inf.cor<-abs(cor(data.exp))
diag(inf.cor)<-0
##convert matrix to edge list
adjmatrix2edgelist(inf.cor)
}

