% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmatrix2edgelist.R
\name{adjmatrix2edgelist}
\alias{adjmatrix2edgelist}
\title{Convert network matrix to edge list}
\usage{
adjmatrix2edgelist(adjmatrix, cutoff = 0, directed = F)
}
\arguments{
\item{adjmatrix}{a network matrix}

\item{cutoff}{threshod to cut the edge weights. Only the edges that pass the threshold will be shown in the output results.}

\item{directed}{logical, if \code{FLASE} the adjmatrix is transformed to symmetric matrix}
}
\value{
A data frame of edge list.
}
\description{
Convert the network matrix to edge list, in which the first column contains the regulators and second column presents the target genes. The edge weights are in the second column.
}
\examples{
##load data
library(RLowPCor)
data(gnwdata)
data.exp<-gnwdata$size10$ts1[,-c(1:2)]
genes<-colnames(data.exp)
##build correlation network
inf.cor<-abs(cor(data.exp))
diag(inf.cor)<-0
##convert matrix to edge list
adjmatrix2edgelist(inf.cor)
}

