% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist2adjmatrix.R
\name{edgelist2adjmatrix}
\alias{edgelist2adjmatrix}
\title{Convert edge list to network matrix}
\usage{
edgelist2adjmatrix(edgelist, genes, cutoff = 0, directed = F)
}
\arguments{
\item{edgelist}{a data frame of edge list of a network, in which columns are regulators, targets and edge weights.}

\item{genes}{gene names to name the rows and columns of the output network matrix.}

\item{cutoff}{the threshold to cut the edge list.}

\item{directed}{logical, to create directed or undirected (symmetric) network matrix.}
}
\value{
\code{edgelist2adjmatrix} returns a network matrix.
}
\description{
The function is to convert edge list to network matrix.
}
\examples{
##load data
library(RLowPC)
data(gnwdata)
data.exp<-gnwdata$size100$ts1[,-c(1:3)]
genes<-colnames(data.exp)
ref.edge<-gnwdata$size100$net1
ref.edge[,3]<-1
ref.adj<-edgelist2adjmatrix(ref.edge,genes,directed=F)
}

