% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowPCor.R
\name{LowPCor}
\alias{LowPCor}
\title{Calculate low order partial correlation}
\usage{
LowPCor(data.exp, p.cut = 0.05, up2 = 2, p.quant.prob = 1,
  method = "pearson", progressbar = T)
}
\arguments{
\item{adjmatrix}{A network matrix}

\item{directed}{Logical. If TRUE, the network is considered as directed. If FALSE, the upper triangular part of the matrix is used to calcuate the rank matrix}
}
\value{
an network connection rank matrix
}
\description{
Calculate up to second order partial correlation (PC).The zero order PC is the general correlation. The first order PC is
calcualted based on zero PC matrix after applying correlation significance testing. In the zero PC matrix, the less significant edges are
removed with a p-value cut-off. The second order PC is calcuated based on first order PC matrix after significance testing. For all the combinations of first order
PC or second order PC of a pair of node, the \code{quantile(p.values,probs=p.quant.prob)} quantile of p-values is selected as the
Such procedures reduce the computational cost, especially in large scale networks. To use fully connected network to calculate the first and second order PC,
users can set the p-value cut-off to \code{p.cut=1.1}.
}

