% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.net.R
\name{random.net}
\alias{random.net}
\title{Resort datasets and generate random networks}
\usage{
random.net(input, nexp = NA, type = "data", directed = F)
}
\arguments{
\item{input}{a gene expression data frame or a string vector of gene names.}

\item{nexp}{a numeric number to indicate how many experiments in the datasets. The gene expression datasets are resorted in both experiment-wise and gene-wise.}

\item{type}{a character string to indicate the resort datasets ("data") or to generate random networks ("network").}

\item{directed}{logical, if \code{FALSE}, the random network generated from gene names will be converted to symmetric matrix.}
}
\value{
\code{random.net} returns a randomly resorted expression data frame or a random network matrix.
}
\description{
The function is used to generate random datasets or random networks
}
\details{
(a) Input a gene expression data frame with genes in columns and samples in rows, the function resorts the sample points for each genes in each experiments.
(b) Input a string vector of gene names, the function generates random network with random edge weightes.
}

