% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.cut.R
\name{table.cut}
\alias{table.cut}
\title{Cut the confusion table at threshods}
\usage{
table.cut(input.table, cutoff = 1000, cutat = "top")
}
\arguments{
\item{input.table}{the confusion table output from \link{table.evalution}.}

\item{cutoff}{the cutoff threshold for top weighted edges, fpr or tpr.}

\item{cutat}{the options are "top", "fpr" and "tpr", which indicate to cut the confusion table at top weighted edges, fpr or tpr.}
}
\value{
a list of results cantains the cut table, the pAUROC value, pAUPR value, precesion for the remained edges, the corresponding top, fpr, tpr cutoffs
and the statistical measures after applying function \link{confusion}.
}
\description{
A top weighted edge, fpr or tpr threshod is used to cut the confusion table output from \link{table.evaluation}. The provided \code{cutoff}.
threshold must match to the scales of the \code{cutat} metrices.
}

