% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.evaluate.R
\name{table.evaluate}
\alias{table.evaluate}
\title{Evaluate inferred network to refence network}
\usage{
table.evaluate(inf.adj, ref.adj, directed = F)
}
\arguments{
\item{inf.adj}{the inferred network matrix. Column names and row names match to the reference network.}

\item{ref.adj}{the reference network matrix with 1 inidating connected edge and 0 unconnected edge.}

\item{directed}{logical, to compare as directed or undirected networks. In a undirected network, only the upper triangular of the network matrix is used for evaluation.}
}
\value{
a confusion table of TP, FP, TN and FN
}
\description{
The inferred network is evaluated by comparing to the reference network. The output is a tables of TP, FP, TN and FP with different edge weight cut-offs
}
\examples{
 library(networkBMA)
 library(RLowPCor)
 ##load DREAM4 size100_1 datasets
 data(dream4)
 data.exp<-dream4ts10[[1]][,-c(1:2)]
 genes<-colnames(data.exp)
 ref.edge<-dream4gold10[[1]]
 ref.adj<-edgelist2adjmatrix(ref.edge,genes)
 inf.cor<-abs(cor(data.exp))
 diag(inf.cor)<-0
 table.cor<-table.evaluate(inf.adj = inf.cor,ref.adj = ref.adj)
 head(table.cor)
##generate statistical measures
 confusion(table.cor)
}
\references{
Meyer PE, Lafitte F, Bontempi G: minet: A R/Bioconductor package for inferring large transcriptional networks using mutual information.
BMC Bioinformatics 2008, 9:461.
}

